+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - replaced [PEP526](https://www.python.org/dev/peps/pep-0526/#abstract)-style variable type hints
     with [PEP484](https://www.python.org/dev/peps/pep-0484/)-compatible
   - fixed `AttributeError` due to unavailable `MagicMock.assert_called_once`
+  - fixed `TypeError` when calling `datetime.datetime.isoformat(datespec=…)`
 
 ## [0.1.0] - 2020-06-16
 ### Added